requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
yfinance>=0.2.0
dimod>=0.12.0
dwave-neal>=0.6.0
quimb>=1.4.0
seaborn>=0.11.0
networkx>=2.6.0